Wrapping in UserSwitcher
We were only using the SecurityDisabler to allow Sinj to update without
a user context. This worked fine for most things, but when removing a
base template from a template this wasn't being taken into account and
an exception was being thrown that crashed the app pool.

Raised with Sitecore support (470391) who suggested this fix.

Default to the sitecore admin user but this can be overridden using an
app setting named "Sinj.RunAsUser".