Allow a user to be specified in scripts where needed
This alters the approach taken in #1 - trying to stop changes to Base
Templates causing unhandled exceptions (Sitecore issue 470391).

Instead of wrapping all requests in a UserSwitcher, allow users to
specify when to switch to a user context in there scripts.

The UserSwitcher adds extra SQL calls to every request, which could
effect performance.

Adds methods/properties to the PushContext that will start and stop the
UserSwitcher. This can then be used in scripts to change the user Sinj
runs as. The user context will stop at the end of the script, unless
endRunAsUser is called first.

Updating the static helper scripts to spot base templates updates, and
throw an error giving details of the fix to the user.